Removed ColorCode.Core
