+++ src/Files.App/ViewModels/UserControls/Previews/CodePreviewViewModel.cs
@@ -3,13 +3,14 @@
 
 using ColorCode;
 using Files.App.ViewModels.Properties;
-using Files.Shared.Helpers;
 using System.Collections.Frozen;
 
 namespace Files.App.ViewModels.Previews
 {
 	public sealed partial class CodePreviewViewModel : BasePreviewModel
 	{
+		private static readonly FrozenDictionary<string, ILanguage> extensions = GetDictionary();
+
 		private string textValue;
 		public string TextValue
 		{
@@ -29,6 +30,9 @@ namespace Files.App.ViewModels.Previews
 		{
 		}
 
+		public static bool IsCodeFile(string extension)
+			=> extensions.ContainsKey(extension);
+
 		public async override Task<List<FileProperty>> LoadPreviewAndDetailsAsync()
 		{
 			var details = new List<FileProperty>();
@@ -38,7 +42,7 @@ namespace Files.App.ViewModels.Previews
 				var text = TextValue ?? await ReadFileAsTextAsync(Item.ItemFile);
 				details.Add(GetFileProperty("PropertyLineCount", text.Split('\n').Length));
 
-				CodeLanguage = FileExtensionHelpers.CodeFileExtensions[Item.FileExtension.ToLowerInvariant()];
+				CodeLanguage = extensions[Item.FileExtension.ToLowerInvariant()];
 				TextValue = text.Left(Constants.PreviewPane.TextCharacterLimit);
 			}
 			catch (Exception e)
@@ -47,6 +51,41 @@ namespace Files.App.ViewModels.Previews
 			}
 
 			return details;
+		}
+
+		private static FrozenDictionary<string, ILanguage> GetDictionary()
+		{
+			var items = new Dictionary<ILanguage, string>
+			{
+				[Languages.Aspx] = "aspx",
+				[Languages.AspxCs] = "acsx",
+				[Languages.Cpp] = "cpp,c++,cc,cp,cxx,h,h++,hh,hpp,hxx,inc,inl,ino,ipp,re,tcc,tpp",
+				[Languages.CSharp] = "cs,cake,csx,linq",
+				[Languages.Css] = "css,scss",
+				[Languages.FSharp] = "fs,fsi,fsx",
+				[Languages.Haskell] = "hs",
+				[Languages.Html] = "razor,cshtml,vbhtml,svelte",
+				[Languages.Java] = "java",
+				[Languages.JavaScript] = "js,jsx",
+				[Languages.Php] = "php",
+				[Languages.PowerShell] = "pwsh,ps1,psd1,psm1",
+				[Languages.Typescript] = "ts,tsx",
+				[Languages.VbDotNet] = "vb,vbs",
+				[Languages.Xml] = "xml,axml,xaml,xsd,xsl,xslt,xlf",
+			};
+
+			var dictionary = new Dictionary<string, ILanguage>();
+
+			foreach (var item in items)
+			{
+				var extensions = item.Value.Split(',').Select(ext => $".{ext}");
+				foreach (var extension in extensions)
+				{
+					dictionary.Add(extension, item.Key);
+				}
+			}
+
+			return dictionary.ToFrozenDictionary();
 		}
 	}
 }
